backend/zetatech_backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-environ==0.11.2
drf-spectacular==0.26.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
Pillow==10.0.0
sqlparse==0.4.4
psycopg2-binary==2.9.6
djoser==2.2.0
drf-extra-fields==3.7.0
django-import-export==3.3.1
django-filter==23.3
python-slugify==8.0.1

redis==5.0.1

pytils==0.4.1
flake8==6.1.0
black==23.9.1
python-dotenv==1.0.0
factory-boy==3.3.0
isort==5.12.0
ruff==0.0.290